+++ examples/breadcrumb/base.jsx
@@ -1,12 +1,13 @@
 import React from 'react';
-import BreadCrumb from '~/components/bread-crumb'; // `~` is replaced with design-system-react at runtime
+import BreadCrumb from '~/components/breadcrumb'; // `~` is replaced with design-system-react at runtime
 
 const Example = React.createClass({
 	displayName: 'BreadCrumbExample',
 
 	render () {
 		const trail = [
-			(<a href="#">Parent Entity</a>)
+			(<a href="javascript:void(0);">Parent Entity</a>),
+			(<a href="javascript:void(0);">Parent Record Name</a>)
 		];
 
 		return (